working on time system - pause - time scale

+++ Assets/Sources/Features/Time/TickComponent.cs
@@ -3,5 +3,5 @@ using Entitas.CodeGenerator;
 
 [Core, SingleEntity]
 public class TickComponent : IComponent {
-	public long currentTick;
+	public float currentTick;
 }
